+++ Assets/Scripts/Buildings/Barracks.cs
@@ -12,7 +12,7 @@ public class Barracks : Building
     public int BarracksIndex;
     public override float BaseCooldown => float.MaxValue / 2;
     public override int BaseDamage => throw new System.NotImplementedException();
-    public override int BaseRange => throw new System.NotImplementedException();
+    public override float BaseRange => throw new System.NotImplementedException();
     public override VerticalRegion AttackRegion => throw new System.NotImplementedException();
 
     private Vector2Int exitGridPosition;
