+++ Assets/Velociraptor.cs
@@ -11,7 +11,7 @@ public class Velociraptor : Enemy
     public override VerticalRegion Region => VerticalRegion.Ground;
     public override float BaseCooldown => AttackSpeed.Fast;
     public override int BaseDamage => 1;
-    public override int BaseRange => 1;
+    public override float BaseRange => 1;
     public override VerticalRegion AttackRegion => VerticalRegion.Ground;
     protected override float BaseMovementSpeed => 2f;
